requirements.txt
Flask
Flask-Session
Flask-SQLAlchemy
python-dotenv
Werkzeug
psycopg2-binary
openai
numpy scipy
matplotlib

